Move examples/cnn_utils to examples/vision
This is more consistent with the examples/language naming scheme. Also
update README to include language example.
