+++ examples/vision/cifar_resnet.py
