+++ examples/vision/datasets.py
@@ -1,4 +1,4 @@
-"""Functions for getting datasets."""
+"""Functions for getting computer vision datasets."""
 from __future__ import annotations
 
 import os
