+++ examples/vision/engine.py
@@ -1,4 +1,4 @@
-"""Train and Eval function."""
+"""Train and Eval functions for computer vision examples."""
 from __future__ import annotations
 
 import argparse
