+++ examples/vision/optimizers.py
@@ -1,4 +1,4 @@
-"""Utilities for getting optimizers."""
+"""Utilities for getting optimizers for computer vision examples."""
 from __future__ import annotations
 
 import argparse
